+++ Gruntfile.js
@@ -16,20 +16,47 @@ module.exports = function (grunt) {
                     console: true,
                     module: true,
                     document: true,
-                    GeneralOmen: true
+                    GeneralOmen: true,
+                    /* MOCHA */
+                    "describe": true,
+                    "it": true,
+                    "before": true,
+                    "beforeEach": true,
+                    "after": true,
+                    "afterEach": true
                 }
             }
         },
-        nodeunit: {
-            all: ['tests/**/*Test.js']
+        mochaTest: {
+            options: {
+                reporter: 'spec',
+                require: [
+                    './engine/utils/general'
+                ]
+            },
+            src: ['tests/**/*Test.js']
+        },
+        mocha_istanbul: {
+            coverage: {
+                options:{
+                    mochaOptions: ['--harmony'], // any extra options
+                    istanbulOptions: ['--harmony'],
+                    require: [
+                        './engine/utils/general'
+                    ]
+                },
+                src: ['tests/*/*Test.js'] // a folder works nicely
+            }
         }
     });
 
     grunt.loadNpmTasks('grunt-contrib-jshint');
-    grunt.loadNpmTasks('grunt-contrib-nodeunit');
+    grunt.loadNpmTasks('grunt-mocha-test');
+    grunt.loadNpmTasks('grunt-mocha-istanbul');
 
-    grunt.registerTask('test', ['jshint', 'nodeunit']);
+    grunt.registerTask('test', ['jshint', 'mochaTest']);
+    grunt.registerTask('coverage', ['mocha_istanbul']);
 
-    grunt.registerTask('default', ['jshint', 'nodeunit']);
+    grunt.registerTask('default', ['jshint', 'mochaTest', 'mocha_istanbul']);
 
 };