+++ tests/base/commandTest.js
@@ -9,49 +9,44 @@
 /*jslint node: true */
 "use strict";
 
-var Command   = require('../../engine/base/command'),
+var assert = require("assert");
+
+var Command = require('../../engine/base/command'),
     uiMockup = require('../../engine/testing/uiMockup');
 
-module.exports = {
-    testEmpty: function (test) {
-        test.expect(4);
+describe("engine.base.command", function () {
+    describe("constructor", function(){
+        it("should be empty", function () {
+            var cmd = new Command();
 
-        var cmd = new Command();
+            assert.equal(cmd.cli, null);
+            assert.equal(cmd.filename, null);
+            assert.equal(cmd.commandName, "");
+            assert.throws(cmd.run, "Command not implemented");
 
-        test.equal(cmd.cli, null);
-        test.equal(cmd.filename, null);
-        test.equal(cmd.commandName, "");
-        test.throws(cmd.run, "Command not implemented");
+        });
+    });
+    describe("init", function(){
+        it("should be empty", function () {
+            var cmd = new Command();
+            cmd.init(null, "");
 
-        test.done();
-    },
-    testEmptyInit: function (test) {
-        test.expect(4);
+            assert.equal(cmd.cli, null);
+            assert.equal(cmd.filename, null);
+            assert.equal(cmd.commandName, "");
+            assert.throws(cmd.run, "Command not implemented");
+        });
+        it("should be initialised", function () {
+            var cmd = new Command("cmd");
+            var cli = new uiMockup();
+            cmd.init(cli, "../packages/simple.json");
 
-        var cmd = new Command();
-        cmd.init(null, "");
-
-        test.equal(cmd.cli, null);
-        test.equal(cmd.filename, null);
-        test.equal(cmd.commandName, "");
-        test.throws(cmd.run, "Command not implemented");
-
-        test.done();
-    },
-    testInit: function (test) {
-        test.expect(6);
-
-        var cmd = new Command("cmd");
-        var cli = new uiMockup();
-        cmd.init(cli, "../packages/simple.json");
-
-        test.notEqual(cmd.cli, null);
-        test.equal(cmd.cli, cli);
-        test.notEqual(cmd.filename, "project.json");
-        test.equal(cmd.filename, "../packages/simple.json");
-        test.equal(cmd.commandName, "cmd");
-        test.throws(cmd.run, "Command not implemented");
-
-        test.done();
-    }
-};
+            assert.notEqual(cmd.cli, null);
+            assert.equal(cmd.cli, cli);
+            assert.notEqual(cmd.filename, "project.json");
+            assert.equal(cmd.filename, "../packages/simple.json");
+            assert.equal(cmd.commandName, "cmd");
+            assert.throws(cmd.run, "Command not implemented");
+        });
+    });
+});