+++ tests/commands/aboutTest.js
@@ -9,106 +9,106 @@
 /*jslint node: true */
 "use strict";
 
-var Command = require('../../engine/commands/about'),
-    uiMockup = require('../../engine/testing/uiMockup');
+var assert = require("assert");
 
-var cli = new uiMockup(),
-    simple = "./tests/packages/simple.json",
+var Command = require('../../engine/commands/about'),
+    uiMockup = require('../../engine/testing/uiMockup'),
+    Exceptions = require("./../../engine/base/exceptions");
+
+var simple = "./tests/packages/simple.json",
     deps = "./tests/packages/deps.json",
     noDeps = "./tests/packages/noDeps.json",
-    cmd;
+    cmd, cli;
 
-module.exports = {
-    setUp: function (callback) {
-        cmd = new Command();
+describe("engine.commands.about", function () {
+    describe("run", function () {
 
-        callback();
-    },
-    testSimpleRun: function (test) {
-        cli.clearAll();
-        cmd.init(cli, simple);
+        before(function () {
+            cmd = new Command();
+            cli = new uiMockup();
+        });
 
-        test.equal(cmd.cli, cli);
-        test.equal(cmd.filename, simple);
-        test.equal(cmd.commandName, "about");
+        beforeEach(function () {
+            cli.clearAll();
+        });
 
-        test.doesNotThrow(function () {
-            cmd.run(simple);
-        }, "Command not implemented");
+        it("simple file", function () {
+            cmd.init(cli, simple);
 
-        var cliLines = cli.getAll();
+            assert.equal(cmd.cli, cli);
+            assert.equal(cmd.filename, simple);
+            assert.equal(cmd.commandName, "about");
 
-        test.equal(cliLines.length, 10);
-        for (var i in cliLines) {
-            if (i < 3)
-                test.equal(cliLines[i].type, "header");
-            else if (i == cliLines.length - 1)
-                test.equal(cliLines[i].type, "end");
-            else
-                test.equal(cliLines[i].type, "ok");
-        }
+            assert.doesNotThrow(function () {
+                cmd.run();
+            }, Exceptions.CommandNotImplemented);
 
-        test.equal(cliLines[3].message, "Name:\t\ttest-simple");
-        test.equal(cliLines[4].message, "Version:\t\t0.1.2");
+            var cliLines = cli.getAll();
 
-        test.done();
-    },
-    testNoDepsRun: function (test) {
-        cli.clearAll();
-        cmd.init(cli, noDeps);
+            assert.equal(cliLines.length, 10);
+            for (var i in cliLines) {
+                if (i < 3)
+                    assert.equal(cliLines[i].type, "header");
+                else if (i == cliLines.length - 1)
+                    assert.equal(cliLines[i].type, "end");
+                else
+                    assert.equal(cliLines[i].type, "ok");
+            }
 
-        test.equal(cmd.cli, cli);
-        test.equal(cmd.filename, noDeps);
-        test.equal(cmd.commandName, "about");
+            assert.equal(cliLines[3].message, "Name:\t\ttest-simple");
+            assert.equal(cliLines[4].message, "Version:\t\t0.1.2");
+        });
+        it("no dependencies", function () {
+            cmd.init(cli, noDeps);
 
-        test.doesNotThrow(function () {
-            cmd.run(noDeps);
-        }, "Command not implemented");
+            assert.equal(cmd.cli, cli);
+            assert.equal(cmd.filename, noDeps);
+            assert.equal(cmd.commandName, "about");
 
-        var cliLines = cli.getAll();
+            assert.doesNotThrow(function () {
+                cmd.run(noDeps);
+            }, Exceptions.CommandNotImplemented);
 
-        test.equal(cliLines.length, 16);
-        for (var i in cliLines) {
-            if (i < 3)
-                test.equal(cliLines[i].type, "header");
-            else if (i == cliLines.length - 1)
-                test.equal(cliLines[i].type, "end");
-            else
-                test.equals(cliLines[i].type == "ok" || cliLines[i].type == "separator", true);
-        }
+            var cliLines = cli.getAll();
 
-        test.equal(cliLines[3].message, "Name:\t\ttest-noDeps");
-        test.equal(cliLines[4].message, "Version:\t\t0.1.1");
+            assert.equal(cliLines.length, 16);
+            for (var i in cliLines) {
+                if (i < 3)
+                    assert.equal(cliLines[i].type, "header");
+                else if (i == cliLines.length - 1)
+                    assert.equal(cliLines[i].type, "end");
+                else
+                    assert.equal(cliLines[i].type == "ok" || cliLines[i].type == "separator", true);
+            }
 
-        test.done();
-    },
-    testDepsRun: function (test) {
-        cli.clearAll();
-        cmd.init(cli, deps);
+            assert.equal(cliLines[3].message, "Name:\t\ttest-noDeps");
+            assert.equal(cliLines[4].message, "Version:\t\t0.1.1");
+        });
+        it("dependencies", function () {
+            cmd.init(cli, deps);
 
-        test.equal(cmd.cli, cli);
-        test.equal(cmd.filename, deps);
-        test.equal(cmd.commandName, "about");
+            assert.equal(cmd.cli, cli);
+            assert.equal(cmd.filename, deps);
+            assert.equal(cmd.commandName, "about");
 
-        test.doesNotThrow(function () {
-            cmd.run(deps);
-        }, "Command not implemented");
+            assert.doesNotThrow(function () {
+                cmd.run(deps);
+            }, Exceptions.CommandNotImplemented);
 
-        var cliLines = cli.getAll();
+            var cliLines = cli.getAll();
 
-        test.equal(cliLines.length, 19);
-        for (var i in cliLines) {
-            if (i < 3)
-                test.equal(cliLines[i].type, "header");
-            else if (i == cliLines.length - 1)
-                test.equal(cliLines[i].type, "end");
-            else
-                test.equals(cliLines[i].type == "ok" || cliLines[i].type == "separator", true);
-        }
+            assert.equal(cliLines.length, 19);
+            for (var i in cliLines) {
+                if (i < 3)
+                    assert.equal(cliLines[i].type, "header");
+                else if (i == cliLines.length - 1)
+                    assert.equal(cliLines[i].type, "end");
+                else
+                    assert.equal(cliLines[i].type == "ok" || cliLines[i].type == "separator", true);
+            }
 
-        test.equal(cliLines[3].message, "Name:\t\ttest-deps");
-        test.equal(cliLines[4].message, "Version:\t\t1.1.1");
-
-        test.done();
-    }
-};
+            assert.equal(cliLines[3].message, "Name:\t\ttest-deps");
+            assert.equal(cliLines[4].message, "Version:\t\t1.1.1");
+        });
+    });
+});