+++ tests/project/extrasTest.js
@@ -8,26 +8,23 @@
  */
 /*jslint node: true */
 "use strict";
+var assert = require("assert");
 
 var Extras = require('./../../engine/project/extras'),
     Exceptions = require('./../../engine/base/exceptions');
 
-module.exports = {
-    contributors: {
-        testOk1: function (test) {
-            test.expect(1);
-            test.doesNotThrow(function () {
+describe("engine.project.extras", function () {
+    describe("contributors", function () {
+        it("one should be ok", function () {
+            assert.doesNotThrow(function () {
                 Extras.contributors([{
                     "name": "Tester Unu",
                     "email": "email@duricu.ro"
                 }]);
             }, Error);
+        });
 
-            test.done();
-        },
-        testOkN: function (test) {
-            test.expect(1);
-
+        it("two or more should be ok", function () {
             var contributors = [
                 {
                     "name": "Tester Unu",
@@ -42,25 +39,19 @@ module.exports = {
                     "email": "third@duricu.ro"
                 }
             ];
-            test.doesNotThrow(function () {
+            assert.doesNotThrow(function () {
                 Extras.contributors(contributors);
             }, Error);
+        });
 
-            test.done();
-        },
-        testEmptyList: function (test) {
-            test.expect(1);
-
-            test.doesNotThrow(function () {
+        it("empty list should be ok", function () {
+            assert.doesNotThrow(function () {
                 Extras.contributors([]);
             }, Error);
+        });
 
-            test.done();
-        },
-        testEmptyName: function (test) {
-            test.expect(1);
-
-            test.throws(function () {
+        it("empty name should not be ok", function () {
+            assert.throws(function () {
                 Extras.contributors([
                     {
                         "name": "",
@@ -68,13 +59,10 @@ module.exports = {
                     }
                 ]);
             }, Exceptions.EmptyValue);
+        });
 
-            test.done();
-        },
-        testEmptyEmail: function (test) {
-            test.expect(1);
-
-            test.throws(function () {
+        it("empty email should not be ok", function () {
+            assert.throws(function () {
                 Extras.contributors([
                     {
                         "name": "Tester",
@@ -82,13 +70,10 @@ module.exports = {
                     }
                 ]);
             }, Exceptions.EmptyValue);
+        });
 
-            test.done();
-        },
-        testEmptyBoth: function (test) {
-            test.expect(1);
-
-            test.throws(function () {
+        it("empty name and email should not be ok", function () {
+            assert.throws(function () {
                 Extras.contributors([
                     {
                         "name": "",
@@ -96,13 +81,10 @@ module.exports = {
                     }
                 ]);
             }, Exceptions.EmptyValue);
+        });
 
-            test.done();
-        },
-        testInvalidName: function (test) {
-            test.expect(1);
-
-            test.throws(function () {
+        it("invalid name should not be ok", function () {
+            assert.throws(function () {
                 Extras.contributors([
                     {
                         "name": "T3Ster App",
@@ -110,27 +92,10 @@ module.exports = {
                     }
                 ]);
             }, Exceptions.InvalidValue);
+        });
 
-            test.done();
-        },
-        testInvalidEmailName: function (test) {
-            test.expect(1);
-
-            test.throws(function () {
-                Extras.contributors([
-                    {
-                        "name": "T3Ster App",
-                        "email": "email###"
-                    }
-                ]);
-            }, Exceptions.InvalidValue);
-
-            test.done();
-        },
-        testInvalidEmail: function (test) {
-            test.expect(1);
-
-            test.throws(function () {
+        it("invalid email should not be ok", function () {
+            assert.throws(function () {
                 Extras.contributors([
                     {
                         "name": "Tester App",
@@ -138,52 +103,20 @@ module.exports = {
                     }
                 ]);
             }, Exceptions.InvalidValue);
+        });
 
-            test.done();
-        },
-        testDuplicateEmail: function (test) {
-            test.expect(1);
+        it("invalid name and email should not be ok", function () {
+            assert.throws(function () {
+                Extras.contributors([
+                    {
+                        "name": "T3Ster App",
+                        "email": "email###"
+                    }
+                ]);
+            }, Exceptions.InvalidValue);
+        });
 
-            var contributors = [
-                {
-                    "name": "Tester Unu",
-                    "email": "email@duricu.ro"
-                },
-                {
-                    "name": "Tester Doi",
-                    "email": "email@duricu.ro"
-                }
-            ];
-
-            test.throws(function () {
-                Extras.contributors(contributors);
-            }, Exceptions.DuplicateValue);
-
-            test.done();
-        },
-        testDuplicateNameEmail: function (test) {
-            test.expect(1);
-
-            var contributors = [
-                {
-                    "name": "Tester Unu",
-                    "email": "email@duricu.ro"
-                },
-                {
-                    "name": "Tester Unu",
-                    "email": "email@duricu.ro"
-                }
-            ];
-
-            test.throws(function () {
-                Extras.contributors(contributors);
-            }, Exceptions.DuplicateValue);
-
-            test.done();
-        },
-        testDuplicateName: function (test) {
-            test.expect(1);
-
+        it("duplicate name should be ok", function () {
             var contributors = [
                 {
                     "name": "Tester Unu",
@@ -195,274 +128,251 @@ module.exports = {
                 }
             ];
 
-            test.doesNotThrow(function () {
+            assert.doesNotThrow(function () {
                 Extras.contributors(contributors);
             }, Error);
+        });
 
-            test.done();
-        },
-        testInvalidParameters: function (test) {
-            test.expect(1);
+        it("duplicate email should not be ok", function () {
+            var contributors = [
+                {
+                    "name": "Tester Unu",
+                    "email": "email@duricu.ro"
+                },
+                {
+                    "name": "Tester Doi",
+                    "email": "email@duricu.ro"
+                }
+            ];
 
-            test.throws(function () {
+            assert.throws(function () {
+                Extras.contributors(contributors);
+            }, Exceptions.DuplicateValue);
+        });
+
+        it("duplicate name and email should not be ok", function () {
+            var contributors = [
+                {
+                    "name": "Tester Unu",
+                    "email": "email@duricu.ro"
+                },
+                {
+                    "name": "Tester Unu",
+                    "email": "email@duricu.ro"
+                }
+            ];
+
+            assert.throws(function () {
+                Extras.contributors(contributors);
+            }, Exceptions.DuplicateValue);
+        });
+
+        it("invalid parameters should not be ok", function () {
+            assert.throws(function () {
                 Extras.contributors("666");
             }, Exceptions.InvalidValue);
+        });
+    });
 
-            test.done();
-        }
-    },
-
-    keywords: {
-        testOk1: function (test) {
-            test.expect(1);
-
-            test.doesNotThrow(function () {
+    describe("keywords", function () {
+        it("one should be ok", function () {
+            assert.doesNotThrow(function () {
                 Extras.keywords(["keyword"]);
             }, Error);
+        });
 
-            test.done();
-        },
-        testOkN: function (test) {
-            test.expect(1);
-
-            test.doesNotThrow(function () {
+        it("two or more should be ok", function () {
+            assert.doesNotThrow(function () {
                 Extras.keywords(["keyword1", "keyword2", "keywordn"]);
             }, Error);
+        });
 
-            test.done();
-        },
-        testEmptyList: function (test) {
-            test.expect(1);
-
-            test.doesNotThrow(function () {
+        it("empty list should be ok", function () {
+            assert.doesNotThrow(function () {
                 Extras.contributors([]);
             }, Exceptions.EmptyValue);
+        });
 
-            test.done();
-        },
-        testEmptyKeyword: function (test) {
-            test.expect(1);
-
-            test.throws(function () {
+        it("empty keywords should not be ok", function () {
+            assert.throws(function () {
                 Extras.keywords([""]);
             }, Exceptions.EmptyValue);
+        });
 
-            test.done();
-        },
-        testInvalidKeyword: function (test) {
-            test.expect(1);
-
-            test.throws(function () {
+        it("invalid keyword should not be ok", function () {
+            assert.throws(function () {
                 Extras.keywords(["k3yw0r#"]);
             }, Exceptions.InvalidValue);
+        });
 
-            test.done();
-        },
-        testDuplicate: function (test) {
-            test.expect(1);
-
-            test.throws(function () {
+        it("duplicate keywords should not be ok", function () {
+            assert.throws(function () {
                 Extras.keywords(["keyword", "keyword"]);
             }, Exceptions.DuplicateValue);
+        });
 
-            test.done();
-        },
-        testInvalidParameters: function (test) {
-            test.expect(1);
-
-            test.throws(function () {
+        it("invalid parameters should not be ok", function () {
+            assert.throws(function () {
                 Extras.keywords(12333);
             }, Exceptions.InvalidValue);
+        });
+    });
 
-            test.done();
-        }
-    },
-
-    homepage: {
-        testOk: function (test) {
-            test.expect(1);
-
-            test.doesNotThrow(function () {
-                Extras.homepage("http://omen.cloud-studio.ro");
-            }, Error);
-
-            test.done();
-        },
-        testEmpty: function (test) {
-            test.expect(2);
-
-            test.throws(function () {
-                Extras.homepage("");
-            }, Exceptions.EmptyValue);
-
-            test.throws(function () {
-                Extras.homepage();
-            }, Exceptions.EmptyValue);
-
-            test.done();
-        },
-        testInvalid: function (test) {
-            test.expect(1);
-
-            test.throws(function () {
-                Extras.homepage("ht##:\\regular");
-            }, Exceptions.InvalidValue);
-
-            test.done();
-        },
-        testInvalidParameters: function (test) {
-            test.expect(1);
-
-            test.throws(function () {
-                Extras.homepage({});
-            }, Exceptions.InvalidValue);
-
-            test.done();
-        }
-    },
-
-    license: {
-        testOkName: function (test) {
-            test.expect(1);
-
-            test.doesNotThrow(function () {
+    describe("homepage", function () {
+        it("name should be ok", function () {
+            assert.doesNotThrow(function () {
                 Extras.license("GPL V.3");
             }, Error);
+        });
 
-            test.done();
-        },
-        testOkUrl: function (test) {
-            test.expect(1);
-
-            test.doesNotThrow(function () {
+        it("url should be ok", function () {
+            assert.doesNotThrow(function () {
                 Extras.license("http://duricu.ro/license.html");
             }, Error);
+        });
 
-            test.done();
-        },
-        testEmpty: function (test) {
-            test.expect(2);
-
-            test.throws(function () {
+        it("empty value should not be ok", function () {
+            assert.throws(function () {
                 Extras.license("");
             }, Exceptions.EmptyValue);
 
-            test.throws(function () {
+            assert.throws(function () {
                 Extras.license();
             }, Exceptions.EmptyValue);
+        });
 
-            test.done();
-        },
-        testInvalid: function (test) {
-            test.expect(1);
-
-            test.throws(function () {
-                Extras.license("ht##:\\regular");
-            }, Exceptions.InvalidValue);
-
-            test.done();
-        },
-        testInvalid2: function (test) {
-            test.expect(1);
-
-            test.throws(function () {
+        it("invalid name should not be ok", function () {
+            assert.throws(function () {
                 Extras.license("G3P3L3 ###");
             }, Exceptions.InvalidValue);
+        });
 
-            test.done();
-        },
-        testInvalidParameters: function (test) {
-            test.expect(1);
+        it("invalid url should not be ok", function () {
+            assert.throws(function () {
+                Extras.license("ht##:\\regular");
+            }, Exceptions.InvalidValue);
+        });
 
-            test.throws(function () {
+        it("invalid parameters should not be ok", function () {
+            assert.throws(function () {
                 Extras.license([]);
             }, Exceptions.InvalidValue);
+        });
+    });
+    describe("license", function () {
+        it("one should be ok", function () {
+            assert.doesNotThrow(function () {
+                Extras.keywords(["keyword"]);
+            }, Error);
+        });
 
-            test.done();
-        }
-    },
+        it("two or more should be ok", function () {
+            assert.doesNotThrow(function () {
+                Extras.keywords(["keyword1", "keyword2", "keywordn"]);
+            }, Error);
+        });
 
-    src: {
-        testOkSource: function (test) {
-            test.expect(5);
+        it("empty list should be ok", function () {
+            assert.doesNotThrow(function () {
+                Extras.contributors([]);
+            }, Exceptions.EmptyValue);
+        });
 
-            test.doesNotThrow(function () {
+        it("empty keywords should not be ok", function () {
+            assert.throws(function () {
+                Extras.keywords([""]);
+            }, Exceptions.EmptyValue);
+        });
+
+        it("invalid keyword should not be ok", function () {
+            assert.throws(function () {
+                Extras.keywords(["k3yw0r#"]);
+            }, Exceptions.InvalidValue);
+        });
+
+        it("duplicate keywords should not be ok", function () {
+            assert.throws(function () {
+                Extras.keywords(["keyword", "keyword"]);
+            }, Exceptions.DuplicateValue);
+        });
+
+        it("invalid parameters should not be ok", function () {
+            assert.throws(function () {
+                Extras.keywords(12333);
+            }, Exceptions.InvalidValue);
+        });
+    });
+    describe("src", function () {
+        it("value should be ok", function () {
+            assert.doesNotThrow(function () {
                 Extras.src("source");
             }, Error);
 
-            test.doesNotThrow(function () {
+            assert.doesNotThrow(function () {
                 Extras.src("src/adm2");
             }, Error);
 
-            test.doesNotThrow(function () {
+            assert.doesNotThrow(function () {
                 Extras.src("src.32_qq/adm.32");
             }, Error);
 
-            test.doesNotThrow(function () {
+            assert.doesNotThrow(function () {
                 Extras.src("_qq/adm.32");
             }, Error);
 
-            test.doesNotThrow(function () {
+            assert.doesNotThrow(function () {
                 Extras.src(".pq");
             }, Error);
+        });
 
-            test.done();
-        },
-        testEmpty: function (test) {
-            test.expect(1);
-
-            test.throws(function () {
+        it("empty value should not be ok", function () {
+            assert.throws(function () {
                 Extras.src("");
             }, Exceptions.EmptyValue);
 
-            test.done();
-        },
-        testInvalid: function (test) {
-            test.expect(6);
+            assert.throws(function () {
+                Extras.src();
+            }, Exceptions.EmptyValue);
+        });
 
-            test.throws(function () {
+        it("invalid value should not be ok", function () {
+            assert.throws(function () {
                 Extras.src("32-aaqqq");
             }, Exceptions.InvalidValue);
 
-            test.throws(function () {
+            assert.throws(function () {
                 Extras.src("#ana#");
             }, Exceptions.InvalidValue);
 
-            test.throws(function () {
+            assert.throws(function () {
                 Extras.src(".correct/a#3");
             }, Exceptions.InvalidValue);
 
-            test.throws(function () {
+            assert.throws(function () {
                 Extras.src("{.correct/a#3}");
             }, Exceptions.InvalidValue);
 
-            test.throws(function () {
+            assert.throws(function () {
                 Extras.src("$122");
             }, Exceptions.InvalidValue);
 
-            test.throws(function () {
+            assert.throws(function () {
                 Extras.src({} + 123);
             }, Exceptions.InvalidValue);
+        });
 
-            test.done();
-        },
-        testInvalidParameters: function (test) {
-            test.expect(3);
-
-            test.throws(function () {
+        it("invalid parameters should not be ok", function () {
+            assert.throws(function () {
                 Extras.src();
             }, Exceptions.InvalidValue);
 
-            test.throws(function () {
+            assert.throws(function () {
                 Extras.src([]);
             }, Exceptions.InvalidValue);
 
-            test.throws(function () {
+            assert.throws(function () {
                 Extras.src({});
             }, Exceptions.InvalidValue);
-
-            test.done();
-        }
-    }
-
-};
+        });
+    });
+});